Real men don't backup, they commit.

+++ babel/babel-get/babel-get.c
@@ -24,7 +24,7 @@
 extern "C" {
 #endif
 int chdir(const char *);
-char *getcwd(char *, int);
+char *getcwd(char *, size_t);
 #ifdef __cplusplus
 }
 #endif
