+++ babel/babel-get/get_dir.c
@@ -20,7 +20,7 @@
 extern "C" {
 #endif
 int chdir(const char *);
-char *getcwd(char *, int);
+char *getcwd(char *, size_t);
 #ifdef __cplusplus
 }
 #endif
